fix: correctly parse structs in Starknet return types
Currently, starknet.js doesn't do a good job with structs, such as the Uint256 struct returned by the ERC20 contract. Instead of returning all the fields of the struct, it just returns the first field.

Concretely, if a user had a balance of 1 in an ERC20 contract, this is what would have been returned from `balance_of` previously:

```json
{ "res": "0x0" }
```

With this change we properly parse the Uint256 struct, so the method will instead return the following:

```json
{ "res": { "low": "0x1", "high": "0x0" } }
```