Update connect_network.md
Include Version in RPC Provider URL to Avoid simulation_flags Error

I've encountered an issue where initializing the RPC provider without specifying the version in the URL causes transactions to fail on the network, resulting in a missing non-optional param: simulation_flags error. This seems to happen because the StarkNet node expects the version to be included in the URL, which is not currently reflected in the documentation.