fix: remove [warning] from typedoc for external usage (#1095)
* test: Improve tests performance (#1121)

* test: fix transaction retry interval fallback for devnet tests

* test: remove test.only

* Update _test.yml (#1126)

* fix: rename the @param to match the definitions

this commit changes the typedoc that do not match the function declaration

* fix: add a space before typedoc <br/> so that it catches the name

when a @param contains an html tag, it should not be attached to the parameter
name otherwise it does consider the tag is part of the name. This commit
detaches the name from the tag

* fix: improve function declaration so that typedoc matches it

this is a specific case when a function has several declarations and the first
one does not match the typedoc because the parameter is optional. This commit
simplifies the declaration by making the comment optional with a ?.

* fix: reintroduce the parameter name

change the declaration of a function so that the parameter name appears.
To do so it moves the assignement in the function.

* fix: remove unused type declaration and export

removes a declaration from account that is not used in the code and
redundant with the exact same one in the provider file. export the
declaration that is the one really used.

---------

Co-authored-by: Luka Saric <32763694+lukasaric@users.noreply.github.com>
Co-authored-by: Ivan Pavičić <ivan.pavicic@live.com>